// apps/web/src/api/model/cardProgramAuthorizationDtoTimeoutDefault.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Allawee Infrastructure Backend APIs
 * The Allawee Infrastructure APIs description
 * OpenAPI spec version: 0.0.1
 */

/**
 * Timeout default action
 */
export type CardProgramAuthorizationDtoTimeoutDefault = typeof CardProgramAuthorizationDtoTimeoutDefault[keyof typeof CardProgramAuthorizationDtoTimeoutDefault];


export const CardProgramAuthorizationDtoTimeoutDefault = {
  approve: 'approve',
  decline: 'decline',
} as const;