// apps/web/src/api/model/cardProgramDistributionDtoType.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Allawee Infrastructure Backend APIs
 * The Allawee Infrastructure APIs description
 * OpenAPI spec version: 0.0.1
 */

export type CardProgramDistributionDtoType = typeof CardProgramDistributionDtoType[keyof typeof CardProgramDistributionDtoType];


export const CardProgramDistributionDtoType = {
  bulked: 'bulked',
  individually: 'individually',
} as const;